requirements.txt
Django==5.2.3
gunicorn
whitenoise[brotli]
dj-database-url
psycopg2-binary
python-dotenv
